Fixed issue 192 (https://github.com/BoltsFramework/Bolts-iOS/issues/192), allowing projects that include frameworks which themselves include Bolts to work without running into “Include of non-modular header inside framework module” compilation errors.
